fix reaction kinetic wrapper & clean up
